+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 ### What needs to be done
 
-* 
+* test inputs before insert them into db
 
 ### What has been done in version 0.0
 
-* 2014-09-
+* 2014-09-11  
+  Can insert records into db  
+  Can view records as a table